Overview placeholder, Started setting up API
I currently made a placeholder for the Overview, and started working on the backend scripts for the database. Currently, there is a script to connect and disconnect from the database, a model for the Posts, and a Posts router that isn't being used at the moment. In the future, I am going to modify the overview to display the posts in the Database.